tidy schema
