+++ pyispyb/core/schemas/robotactions.py
@@ -1,0 +1,12 @@
+from typing import Optional
+
+from pydantic import BaseModel
+
+
+class RobotAction(BaseModel):
+    actionType: str
+    status: Optional[str]
+    message: Optional[str]
+
+    class Config:
+        orm_mode = True
